Update default keymap to play/pause media on Layer(1)
Hey Dennis,

I was playing around with my bento that I got from you and I found it useful to have a different knob press functionality on the second layer in form of play/pause media and so thought I'd share it with you with this PR. 

I tested this with my bento and works fine. 

Let me know what you think?